refactor: split the modules this work grew past the readability limit
Three files crossed a readability threshold as the session-scoped read and
the ATIF exporter landed. Splitting them along an existing seam is the fix;
recording the growth in the ratchet baseline would only bank the debt.

- `temper-ots/src/atif.rs` becomes `atif/mod.rs` (the mapping) and
  `atif/types.rs` (the ATIF document schema). The mapping tables stay with
  the mapping; the schema types stand alone.
- `temper-store-turso/src/store/trajectory.rs` keeps the write path and hands
  the read path — recent rows, unmet intents, stats, agent- and
  session-scoped replay — to `trajectory_queries.rs`.
- The `ObserveReadStore` impls for Postgres and Turso move out of
  `temper-server/src/storage/mod.rs` into `storage/observe_read.rs`,
  continuing the extraction that took trajectory row encoding out of that
  module. The composite-trait module drops from 2847 to 2707 lines and stops
  being the workspace's largest file by a wider margin.

No behavior change: the moved code is unmodified.
