+++ crates/temper-server/src/storage/mod.rs
@@ -34,6 +34,7 @@ use crate::platform_store::PlatformStore;
 use crate::platform_store::SimPlatformStore;
 use crate::state::trajectory::TrajectoryEntry;
 
+mod observe_read;
 mod published_artifacts;
 mod query_plane_impls;
 mod query_plane_read;
@@ -1753,147 +1754,6 @@ impl BackendNamedStore for PostgresEventStore {
 impl BackendNamedStore for TursoEventStore {
     fn backend_name(&self) -> &'static str {
         "turso"
-    }
-}
-
-#[async_trait::async_trait]
-impl ObserveReadStore for PostgresEventStore {
-    async fn load_recent_trajectories(
-        &self,
-        limit: i64,
-    ) -> Result<Vec<TursoTrajectoryRow>, PersistenceError> {
-        self.load_recent_trajectories(limit)
-            .await
-            .map(|rows| rows.into_iter().map(pg_trajectory_to_turso).collect())
-    }
-
-    async fn load_unmet_intent_rows(&self) -> Result<Vec<UnmetIntentAggRow>, PersistenceError> {
-        self.load_unmet_intent_rows()
-            .await
-            .map(|rows| rows.into_iter().map(pg_unmet_to_turso).collect())
-    }
-
-    async fn load_submit_spec_timestamps(
-        &self,
-    ) -> Result<BTreeMap<String, String>, PersistenceError> {
-        self.load_submit_spec_timestamps().await
-    }
-
-    async fn count_trajectories_by_tenant(
-        &self,
-    ) -> Result<BTreeMap<String, u64>, PersistenceError> {
-        self.count_trajectories_by_tenant().await
-    }
-
-    async fn query_trajectory_stats(
-        &self,
-        entity_type: Option<&str>,
-        action: Option<&str>,
-        success_filter: Option<bool>,
-        failed_limit: i64,
-    ) -> Result<TrajectoryStats, PersistenceError> {
-        self.query_trajectory_stats(entity_type, action, success_filter, failed_limit)
-            .await
-            .map(pg_stats_to_turso)
-    }
-
-    async fn query_trajectories_by_agent(
-        &self,
-        agent_id: &str,
-        tenant: Option<&str>,
-        entity_type: Option<&str>,
-        limit: i64,
-    ) -> Result<Vec<TursoTrajectoryRow>, PersistenceError> {
-        self.query_trajectories_by_agent(agent_id, tenant, entity_type, limit)
-            .await
-            .map(|rows| rows.into_iter().map(pg_trajectory_to_turso).collect())
-    }
-
-    async fn query_trajectories_by_session(
-        &self,
-        session_id: &str,
-        tenant: Option<&str>,
-        entity_type: Option<&str>,
-        limit: i64,
-    ) -> Result<Vec<TursoTrajectoryRow>, PersistenceError> {
-        self.query_trajectories_by_session(session_id, tenant, entity_type, limit)
-            .await
-            .map(|rows| rows.into_iter().map(pg_trajectory_to_turso).collect())
-    }
-
-    async fn query_agent_summaries(
-        &self,
-        tenant: Option<&str>,
-    ) -> Result<Vec<AgentSummary>, PersistenceError> {
-        self.query_agent_summaries(tenant)
-            .await
-            .map(|rows| rows.into_iter().map(pg_agent_summary_to_turso).collect())
-    }
-}
-
-#[async_trait::async_trait]
-impl ObserveReadStore for TursoEventStore {
-    async fn load_recent_trajectories(
-        &self,
-        limit: i64,
-    ) -> Result<Vec<TursoTrajectoryRow>, PersistenceError> {
-        self.load_recent_trajectories(limit).await
-    }
-
-    async fn load_unmet_intent_rows(&self) -> Result<Vec<UnmetIntentAggRow>, PersistenceError> {
-        self.load_unmet_intent_rows().await
-    }
-
-    async fn load_submit_spec_timestamps(
-        &self,
-    ) -> Result<BTreeMap<String, String>, PersistenceError> {
-        self.load_submit_spec_timestamps().await
-    }
-
-    async fn count_trajectories_by_tenant(
-        &self,
-    ) -> Result<BTreeMap<String, u64>, PersistenceError> {
-        self.count_trajectories_by_tenant().await
-    }
-
-    async fn query_trajectory_stats(
-        &self,
-        entity_type: Option<&str>,
-        action: Option<&str>,
-        success_filter: Option<bool>,
-        failed_limit: i64,
-    ) -> Result<TrajectoryStats, PersistenceError> {
-        self.query_trajectory_stats(entity_type, action, success_filter, failed_limit)
-            .await
-    }
-
-    async fn query_trajectories_by_agent(
-        &self,
-        agent_id: &str,
-        tenant: Option<&str>,
-        entity_type: Option<&str>,
-        limit: i64,
-    ) -> Result<Vec<TursoTrajectoryRow>, PersistenceError> {
-        self.query_trajectories_by_agent(agent_id, tenant, entity_type, limit)
-            .await
-    }
-
-    async fn query_trajectories_by_session(
-        &self,
-        session_id: &str,
-        tenant: Option<&str>,
-        entity_type: Option<&str>,
-        limit: i64,
-    ) -> Result<Vec<TursoTrajectoryRow>, PersistenceError> {
-        self.query_trajectories_by_session(session_id, tenant, entity_type, limit)
-            .await
-    }
-
-    async fn query_agent_summaries(
-        &self,
-        tenant: Option<&str>,
-    ) -> Result<Vec<AgentSummary>, PersistenceError> {
-        self.query_agent_summaries(tenant).await
     }
 }
